blockchain/Eldorion [Very Easy]/htb/requirements.txt
pwntools
web3
py-solc-x
